requirements.txt
Django
psycopg2-binary
pillow
ipython
djangorestframework-simplejwt
djangorestframework
drf-yasg